refactor(admin): split flags route panels
Why:
- keep the Quantercise admin routes on the same page/content/sections structure
- separate feature flag data loading from summary and grouped row rendering

Verify:
- pnpm exec prettier --check apps/admin/src/app/quantercise/flags/page.tsx apps/admin/src/app/quantercise/flags/flags-content.tsx apps/admin/src/app/quantercise/flags/flags-sections.tsx
- git diff --check -- apps/admin/src/app/quantercise/flags/page.tsx apps/admin/src/app/quantercise/flags/flags-content.tsx apps/admin/src/app/quantercise/flags/flags-sections.tsx
- pnpm --filter @anipotts/admin typecheck
- pnpm --filter @anipotts/admin lint
- pnpm --filter @anipotts/admin build
- pnpm validate
